[MOV] web,*: move LazyComponent
This commit moves the `LazyComponent` class to its own file, to remove
the dependency to `@web/core/registry` from the `@web/core/assets`
module.

This change has been done to allow the 'assets' module to be cached and
re-used in tests in the future, which is not possible if it holds a
retainer to registries.
