+++ addons/website/static/src/client_actions/website_preview/website_builder_action.js
@@ -1,14 +1,8 @@
 import { useComponent, useLayoutEffect, useRef, useState, useSubEnv } from "@web/owl2/utils";
 import { LocalOverlayContainer } from "@html_editor/local_overlay_container";
-import {
-    Component,
-    onMounted,
-    onWillDestroy,
-    onWillStart,
-    onWillUnmount,
-    status,
-} from "@odoo/owl";
-import { LazyComponent, loadBundle } from "@web/core/assets";
+import { Component, onMounted, onWillDestroy, onWillStart, onWillUnmount, status } from "@odoo/owl";
+import { loadBundle } from "@web/core/assets";
+import { LazyComponent } from "@web/core/lazy_component";
 import { browser } from "@web/core/browser/browser";
 import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
 import { _t } from "@web/core/l10n/translation";
